+++ inclide/ConsoleInitializer.h
@@ -1,0 +1,9 @@
+#pragma once
+
+/*
+ * Platform-specific console initialization
+ * Encapsulates all Windows-specific code in a dedicated namespace
+ */
+namespace ConsoleInitializer {
+    void initialize();
+}
